requirements.txt

fastapi==0.108.0
uvicorn==0.27.0
pydantic==2.5.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
pdf2image==1.17.0
pytesseract==0.3.10
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
tenacity==8.2.3
loguru==0.7.2
flake8==6.1.0
mypy==1.7.1
black==23.11.0
